添加了小车开始停止的逻辑函数

+++ run/APP/record.h
@@ -8,8 +8,8 @@ typedef struct
     char txt[25];
     char txt2[25];
     char txt3[25];
-    uint8_t ch1;
-    uint8_t ch2;
+    uint16_t ch1;
+    uint16_t ch2;
     uint16_t x1;
     uint16_t x2;
     uint16_t max1;
@@ -17,6 +17,7 @@ typedef struct
     uint16_t max2;
     uint16_t mid2;
 } record_t;
+
 
 
 extern record_t co60_record;
